Changed weight from float to int

+++ include/helpers.h
@@ -2,4 +2,4 @@
 
 Node* generate();
 
-float treeval(Node *root);
+int treeval(Node *root);
